requirements.txt
python-dotenv
pytelegrambotapi
# pillow
redis